web/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
deprecation==2.1.0
distro==1.9.0
fastapi==0.115.12
Flask==3.1.1
flask-cors==5.0.1
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
multidict==6.4.3
narwhals==1.39.1
numpy==2.2.6
openai==1.78.1
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pluggy==1.6.0
postgrest==1.0.1
propcache==0.3.1
protobuf==6.31.0
pyarrow==20.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
PyJWT==2.10.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.3
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
storage3==0.11.3
streamlit==1.45.1
StrEnum==0.4.15
supabase==2.15.1
supafunc==0.9.4
tenacity==9.1.2
toml==0.10.2
tornado==6.5
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.5
websockets==14.2
Werkzeug==3.1.3
yarl==1.20.0
yt-dlp==2025.4.30
